Added arrays to my code instead of constantly adding objects to create new flies
also I made so that each type the fly reset it becomes a new type of fly either freeze fly, poison fly, or a regular fly. The color of the fly also changes based on the type of fly it is. The type of fly gets stored into a fly name variable. Next I want to create an array of different numbers that control the probability of the fly being freezing, poisonous or regular