Added the pipes to the game
I made sure the pipes don't overlap when they spawn randomly and I also made sure the top and bottom pipes have an equal gap in the middle. Next, I am going to add the bird and pipe collision that restarts the game. Also I want to ensure the pipes only start moving when the player clicks to start the game.