Added a score system that adds points each time the bird goes past a pipe
Next I will start the base code for the second variation of flappy bird (Bird post). In this variation the bird still goes thru pipes but uses the space bar to drop letters to other bird families. The bird families will appear at random moments at the bottom of the screen and move to the left like the pipes. If the bird doesn't deliver the mail to 3 families then the player loses.