Started the bird boss variation.
Created a new js file to store the bird boss variables and drew the boss as well. Next I will code the boss to shoot a red laser beam out of its eyes if the bird at its eye level. But if the bird is lower than its eye level then the boss's mouth will shoot out the laser. I want to create 3 states for the boss. lockOn, laser, coolDown. The lockOn state consists of the boss look for a position to lockOnto to shoot the laser, the laser shoots out the laser and the coolDown is the rest time before the cycle repeats.