I made lasers shoot out from the boss's eyes using the states I mentioned earlier!
I ended up not having a cool down state and just made the laser fade out instead.  I will make the mouth to shoot out lasers next depending on if the bird is in the lower half of the screen. The thing is I don't know if I should make the mouth shoot out a big laser that covers almost the whole bottom half, or if the mouth should act like the eye and follow the bird.